[test] TariffTest: combine unit test classes
